# ai-service/openapi.json
{
  "components": {
    "schemas": {
...
      "EvaluateExplanationRequest": {
        "properties": {
          "audioBase64": {
            "description": "Base64-encoded audio bytes.",
            "title": "Audiobase64",
            "type": "string"
          },
          "context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "",
            "description": "Optional reference context.",
            "title": "Context"
          },
          "filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Defaults to recording.<ext> for mimeType.",
            "title": "Filename"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional BCP-47 language hint, e.g. en.",
            "title": "Language"
          },
          "mimeType": {
            "default": "audio/webm",
            "title": "Mimetype",
            "type": "string"
          },
          "question": {
            "description": "Interview question being answered.",
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "audioBase64",
          "question"
        ],
        "title": "EvaluateExplanationRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IngestResumeRequest": {
        "properties": {
          "content_base64": {
            "minLength": 1,
            "title": "Content Base64",
            "type": "string"
          },
          "filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filename"
          },
          "resume_id": {
            "minLength": 1,
            "title": "Resume Id",
            "type": "string"
          },
          "user_id": {
            "minLength": 1,
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "resume_id",
          "content_base64"
        ],
        "title": "IngestResumeRequest",
        "type": "object"
      },
...
      "ResumeIngestResponse": {
        "properties": {
          "charCount": {
            "title": "Charcount",
            "type": "integer"
          },
          "chunkCount": {
            "title": "Chunkcount",
            "type": "integer"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "namespace": {
            "title": "Namespace",
            "type": "string"
          },
          "pageCount": {
            "title": "Pagecount",
            "type": "integer"
          },
          "parsedSections": {
            "items": {
              "type": "string"
            },
            "title": "Parsedsections",
            "type": "array"
          },
          "sections": {
            "items": {
              "type": "string"
            },
            "title": "Sections",
            "type": "array"
          },
          "userId": {
            "title": "Userid",
            "type": "string"
          }
        },
        "required": [
          "userId",
          "namespace",
          "chunkCount",
          "sections",
          "pageCount",
          "charCount",
          "parsedSections",
          "filename"
        ],
        "title": "ResumeIngestResponse",
        "type": "object"
      },
      "ResumePurgeResponse": {
        "properties": {
          "deletedChunks": {
            "title": "Deletedchunks",
            "type": "integer"
          },
          "namespace": {
            "title": "Namespace",
            "type": "string"
          },
          "namespaceDropped": {
            "title": "Namespacedropped",
            "type": "boolean"
          },
          "remainingChunks": {
            "title": "Remainingchunks",
            "type": "integer"
          },
          "verified": {
            "title": "Verified",
            "type": "boolean"
          }
        },
        "required": [
          "namespace",
          "deletedChunks",
          "namespaceDropped",
          "remainingChunks",
          "verified"
        ],
        "title": "ResumePurgeResponse",
        "type": "object"
      },
...
      "RetrieveResumeRequest": {
        "properties": {
          "company": {
            "default": "amazon",
            "title": "Company",
            "type": "string"
          },
          "stage": {
            "default": "behavioral",
            "title": "Stage",
            "type": "string"
          },
          "top_k": {
            "default": 4,
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Top K",
            "type": "integer"
          },
          "user_id": {
            "minLength": 1,
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id"
        ],
        "title": "RetrieveResumeRequest",
        "type": "object"
      },
      "RubricSectionScore": {
        "properties": {
          "notes": {
            "description": "Short rationale for the section score.",
            "title": "Notes",
            "type": "string"
          },
          "score": {
            "description": "Section score from 1 to 10.",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Score",
            "type": "integer"
          }
        },
        "required": [
          "score",
          "notes"
        ],
        "title": "RubricSectionScore",
        "type": "object"
      },
...
      "TranscribeRequest": {
        "properties": {
          "audioBase64": {
            "description": "Base64-encoded audio bytes.",
            "title": "Audiobase64",
            "type": "string"
          },
          "filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Defaults to recording.<ext> for mimeType.",
            "title": "Filename"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional BCP-47 language hint, e.g. en.",
            "title": "Language"
          },
          "mimeType": {
            "default": "audio/webm",
            "title": "Mimetype",
            "type": "string"
          }
        },
        "required": [
          "audioBase64"
        ],
        "title": "TranscribeRequest",
        "type": "object"
      },
      "TranscriptResponse": {
        "properties": {
          "transcript": {
            "title": "Transcript",
            "type": "string"
          }
        },
        "required": [
          "transcript"
        ],
        "title": "TranscriptResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VoiceEvaluationResponse": {
        "properties": {
          "evaluation": {
            "$ref": "#/components/schemas/VoiceRubricOutput"
          },
          "transcript": {
            "title": "Transcript",
            "type": "string"
          }
        },
        "required": [
          "transcript",
          "evaluation"
        ],
        "title": "VoiceEvaluationResponse",
        "type": "object"
      },
      "VoiceRubricOutput": {
        "properties": {
          "communicationClarity": {
            "$ref": "#/components/schemas/RubricSectionScore"
          },
          "completeness": {
            "$ref": "#/components/schemas/RubricSectionScore"
          },
          "overallScore": {
            "description": "Overall score from 1 to 10.",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Overallscore",
            "type": "integer"
          },
          "strengths": {
            "items": {
              "type": "string"
            },
            "title": "Strengths",
            "type": "array"
          },
          "suggestions": {
            "items": {
              "type": "string"
            },
            "title": "Suggestions",
            "type": "array"
          },
          "technicalCorrectness": {
            "$ref": "#/components/schemas/RubricSectionScore"
          },
          "weaknesses": {
            "items": {
              "type": "string"
            },
            "title": "Weaknesses",
            "type": "array"
          }
        },
        "required": [
          "overallScore",
          "technicalCorrectness",
          "communicationClarity",
          "completeness",
          "strengths",
          "weaknesses",
          "suggestions"
        ],
        "title": "VoiceRubricOutput",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "InterviewForge AI Service",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/resume/ingest": {
      "post": {
        "operationId": "ingest_resume_api_resume_ingest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestResumeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeIngestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest Resume",
        "tags": [
          "resume"
        ]
      }
    },
    "/api/resume/retrieve": {
      "post": {
        "operationId": "retrieve_resume_api_resume_retrieve_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetrieveResumeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retrieve Resume",
        "tags": [
          "resume"
        ]
      }
    },
    "/api/resume/{user_id}": {
      "delete": {
        "operationId": "delete_resume_api_resume__user_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumePurgeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Resume",
        "tags": [
          "resume"
        ]
      }
    },
    "/api/resume/{user_id}/stats": {
      "get": {
        "operationId": "resume_stats_api_resume__user_id__stats_get",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resume Stats",
        "tags": [
          "resume"
        ]
      }
    },
    "/api/speech/evaluate-explanation": {
      "post": {
        "operationId": "evaluate_explanation_api_speech_evaluate_explanation_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluateExplanationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceEvaluationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Evaluate Explanation",
        "tags": [
          "speech"
        ]
      }
    },
    "/api/speech/transcribe": {
      "post": {
        "operationId": "transcribe_api_speech_transcribe_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TranscribeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranscriptResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Transcribe",
        "tags": [
          "speech"
        ]
      }
    },
...
  }
}